Cast to int to avoid deprecation warning on PHP 8.1

+++ src/Parser.php
@@ -71,12 +71,12 @@ class Parser
 
         $wkb12 = false;
 
-        if (($type / 1000) & 1) {
+        if (((int) ($type / 1000)) & 1) {
             $hasZ = true;
             $wkb12 = true;
         }
 
-        if (($type / 1000) & 2) {
+        if (((int) ($type / 1000)) & 2) {
             $hasM = true;
             $wkb12 = true;
         }
